Minor keybindings for math console and clean up remainder of file.
Jupyter QtConsole has a bug which prevents the terminal from scrolling with
graphical output. (see https://github.com/jupyter/qtconsole/issues/386) The
quick fix is to send a clear character to the terminal to reset the position to
the top manually.
Maybe there is a way to scroll instead and append that to the
send lines functions? (scroll all the way down and up a little)